LayoutTests/imported/w3c/web-platform-tests/html/browsers/the-window-object/garbage-collection-and-browsing-contexts/discard_iframe_history_4-expected.txt


FAIL Removing iframe from document removes it from history assert_array_equals: expected property 2 to be 1 but got 2 (expected array [1, 2, 1] got [1, 2, 2])
